% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfp_phosphate.R
\name{rfp_p_gld}
\alias{rfp_p_gld}
\title{Calculate recommended P dose for grassland in the Netherlands}
\usage{
rfp_p_gld(B_LU_CAT, B_SOILTYPE_USDA, A_P_AL, A_P_CC)
}
\arguments{
\item{B_LU_CAT}{(integer) The crop category. Options: arable, grassland, nature}

\item{B_SOILTYPE_USDA}{(character) The  type of soil. Options: clay, silty clay, loam, sand}

\item{A_P_AL}{(numeric) The P-content, measured via ammonium lactate extraction (mg P2O5 / 100g soil)}

\item{A_P_CC}{(numeric) The P-content, measured via calcium chloride extraction (mg P/ kg soil)}
}
\description{
This function calculates the recommended P dose (kg P2O5 / ha) given P-status of the soil.
}
\references{
none yet
}
